+++ lib/Integrations/ElasticPress.php
@@ -9,14 +9,14 @@ use WP_Query;
 /**
  * Hooks for integrating with ElasticPress.
  *
- * @since Unreleased
+ * @since 0.4.3
  */
 class ElasticPress {
 
 	/**
 	 * Register hooks.
 	 *
-	 * @since Unreleased
+	 * @since 0.4.3
 	 */
 	public function register() {
 		if ( ! \is_plugin_active( 'elasticpress/elasticpress.php' ) ) {
@@ -31,7 +31,7 @@ class ElasticPress {
 	/**
 	 * Add argument to ElasticPress's sync query arguments to not apply Unbabble's lang filter.
 	 *
-	 * @since Unreleased
+	 * @since 0.4.3
 	 *
 	 * @param array $args
 	 * @return array
@@ -45,7 +45,7 @@ class ElasticPress {
 	 * Add post language to meta entries sent to ElasticPress in order to be able to filter by
 	 * language later.
 	 *
-	 * @since Unreleased
+	 * @since 0.4.3
 	 *
 	 * @param array   $meta
 	 * @param WP_Post $post
@@ -69,7 +69,7 @@ class ElasticPress {
 	 * Using the `ep_skip_query_integration` filter since its the last filter we can use to
 	 * specifically change the fetch query, and there is no appropriate filter otherwise.
 	 *
-	 * @since Unreleased
+	 * @since 0.4.3
 	 *
 	 * @param bool     $return
 	 * @param WP_Query $query